Cleaned up files and added functionality for transcript
